Intergrating mainscreen window
